pyproject.toml
[project]
name = "2026sp-420-notavirus-exe"
version = "0.1.0"
description = "Feature extension for Scheduler"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "course-constraint-scheduler>=2.6.1",
]

[dependency-groups]
dev = [
    "iniconfig>=2.3.0",
    "packaging>=26.0",
    "pluggy>=1.6.0",
    "pygments>=2.19.2",
    "pytest>=9.0.2",
]
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
